% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, thin)
}
\arguments{
\item{N}{Integer, the number of samples to generate.}

\item{thin}{Integer, the number of iterations between each sample.}
}
\value{
A matrix with N rows and 2 columns. Each row contains a sample of the two variables.
}
\description{
This function performs Gibbs sampling using Rcpp for efficient computation.
}
\examples{
\dontrun{
# Perform Gibbs sampling
samples <- gibbsC(100, 10)

# Inspect the first few samples
print(head(samples))

# Plot the sampled variables
par(mfrow = c(2, 1))
plot(samples[, 1], type = "l", main = "Sampled X", xlab = "Iteration", ylab = "X")
plot(samples[, 2], type = "l", main = "Sampled Y", xlab = "Iteration", ylab = "Y")
}
}
