% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ sparse_matrix_threshold_cpp.R
\name{sparse_matrix_threshold_cpp}
\alias{sparse_matrix_threshold_cpp}
\title{Sparse Matrix Thresholding using Rcpp}
\usage{
sparse_matrix_threshold_cpp(mat, threshold)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{threshold}{A numeric value. Entries in the matrix less than this value will be set to zero.}
}
\value{
A numeric matrix with values below the threshold set to zero.
}
\description{
Apply a threshold to a numeric matrix: values below the threshold are set to zero.
}
\examples{
mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
sparse_mat <- sparse_matrix_threshold_cpp(mat, 0.5)
}
